refactor(nn): split model.rs into specs, predictor, network
